requirements.txt
tweepy>=4.14.0
pytesseract>=0.3.10
Pillow>=10.0.0
requests>=2.31.0
